Refactor `errors` into its own package to allow `NoRetryError` in `step`
