+++ errors/errors_test.go
@@ -1,4 +1,4 @@
-package inngestgo
+package errors
 
 import (
 	"fmt"
@@ -10,13 +10,13 @@ import (
 
 func TestIsNoRetryError(t *testing.T) {
 	err := fmt.Errorf("error")
-	require.False(t, isNoRetryError(err))
+	require.False(t, IsNoRetryError(err))
 
 	wrapped := NoRetryError(err)
-	require.True(t, isNoRetryError(wrapped))
+	require.True(t, IsNoRetryError(wrapped))
 
 	cause := fmt.Errorf("error: %w", wrapped)
-	require.True(t, isNoRetryError(cause))
+	require.True(t, IsNoRetryError(cause))
 }
 
 func TestGetRetryAtTime(t *testing.T) {
@@ -26,15 +26,15 @@ func TestGetRetryAtTime(t *testing.T) {
 	at := RetryAtError(err, expected)
 
 	t.Run("It returns the time with a RetryAtError", func(t *testing.T) {
-		require.NotNil(t, getRetryAtTime(at))
-		require.EqualValues(t, expected, *getRetryAtTime(at))
+		require.NotNil(t, GetRetryAtTime(at))
+		require.EqualValues(t, expected, *GetRetryAtTime(at))
 	})
 
 	t.Run("It returns if RetryAtError is wrapped itself", func(t *testing.T) {
 
 		wrapped := fmt.Errorf("wrap: %w", at)
 
-		require.NotNil(t, getRetryAtTime(wrapped))
-		require.EqualValues(t, expected, *getRetryAtTime(wrapped))
+		require.NotNil(t, GetRetryAtTime(wrapped))
+		require.EqualValues(t, expected, *GetRetryAtTime(wrapped))
 	})
 }
